// src/app/models/procapacitacion.ts
export class Procapacitacion {
    id: number;
    codempresa: string;
    sede: string;
    codtema: number;
    codtipocap: number;
    codmtvcap: number;
    fechaprogram: string;
    fechavigencia: string;
    fechalimite: string;
    codestado: string;
    nrohoras: number;
    showCargaFile: boolean;
    programados: number;
    asistieron: number;
    noasistieron: number;
    codempresares: string;
    codpersonalres: string;
    nombreresp: string;
}